Fixed error, Inital Agent positions, Dynamic Density
Removed an if statement that caused an error with the indexing of a matrix.

Decided to make the velocity function not depend directly on the amount of mass the agent detects. This is because the agents move towards the centriods of their assigned voronai regions (point near the majority of the mass of the agent's region).

Fiddled with the dynamic density function. Not sure how to go about updating density values near agent